Add files via upload
The code above contains a set of SQL queries written in T-SQL, used to retrieve and manipulate data in a database. The queries cover a range of topics, such as joining tables, calculating aggregate functions, filtering data, and creating temporary tables. The code can be used as a reference for developers and data analysts working with T-SQL and SQL Server.